refactor(nexus-web): migrate account notifications to semantic CSS
Split notifications into a dedicated module with semantic classes, wire the route via re-export, and enable the nav item so settings match other account pages.
