+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/account/notifications/notifications.css
@@ -1,0 +1,172 @@
+/* account/notifications — semantic page styles (vanilla CSS, theme tokens) */
+
+.account-notifications-page {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-8);
+}
+
+/* Header */
+.account-notifications-header {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
+}
+
+.account-notifications-title {
+  margin: 0;
+  font-size: var(--font-size-lg);
+  font-weight: var(--font-weight-semibold);
+  line-height: var(--line-height-lg);
+  color: var(--foreground-hex);
+}
+
+.account-notifications-subtitle {
+  margin: 0;
+  font-size: var(--font-size-sm);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+}
+
+/* Settings table */
+.account-notifications-table-card {
+  border: 1px solid var(--border-hex);
+  border-radius: var(--org-border-radius, 0px);
+  background-color: var(--card-hex);
+  overflow: hidden;
+}
+
+.account-notifications-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.account-notifications-table-head-row {
+  border-bottom: 1px solid var(--border-hex);
+  text-align: left;
+  font-size: var(--font-size-sm);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+}
+
+.account-notifications-table-head-cell {
+  padding: var(--space-4);
+  font-weight: var(--font-weight-medium);
+}
+
+.account-notifications-table-head-cell-center {
+  text-align: center;
+}
+
+.account-notifications-column-heading {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--space-2);
+}
+
+.account-notifications-table-row {
+  border-bottom: 1px solid var(--border-hex);
+}
+
+.account-notifications-table-row:last-child {
+  border-bottom: none;
+}
+
+.account-notifications-table-cell {
+  padding: var(--space-4);
+}
+
+.account-notifications-table-cell-center {
+  text-align: center;
+}
+
+.account-notifications-setting-label {
+  margin: 0;
+  font-weight: var(--font-weight-medium);
+  color: var(--foreground-hex);
+}
+
+.account-notifications-setting-description {
+  margin: 0;
+  font-size: var(--font-size-sm);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+}
+
+/* Toggle switches */
+.account-notifications-toggle {
+  position: relative;
+  display: inline-flex;
+  align-items: center;
+  width: 44px;
+  height: 24px;
+  padding: 0;
+  border: none;
+  border-radius: 9999px;
+  cursor: pointer;
+  transition: background-color 0.15s ease;
+}
+
+.account-notifications-toggle-on {
+  background-color: var(--color-primary);
+}
+
+.account-notifications-toggle-off {
+  background-color: var(--muted-bg-hex);
+}
+
+.account-notifications-toggle-knob {
+  position: absolute;
+  top: 2px;
+  left: 2px;
+  width: 20px;
+  height: 20px;
+  border-radius: 9999px;
+  background-color: var(--color-primary-foreground);
+  transition: transform 0.15s ease;
+}
+
+.account-notifications-toggle-on .account-notifications-toggle-knob {
+  transform: translateX(20px);
+}
+
+.account-notifications-toggle-off .account-notifications-toggle-knob {
+  transform: translateX(0);
+}
+
+/* Save button */
+.account-notifications-actions {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.account-notifications-save-button {
+  display: flex;
+  align-items: center;
+  gap: var(--space-2);
+  padding: var(--space-2) var(--space-4);
+  border: none;
+  border-radius: var(--org-border-radius, 0px);
+  background-color: var(--color-primary);
+  font-size: var(--font-size-sm);
+  font-weight: var(--font-weight-medium);
+  line-height: var(--line-height-sm);
+  color: var(--color-primary-foreground);
+  cursor: pointer;
+  transition: background-color 0.15s ease, color 0.15s ease;
+}
+
+.account-notifications-save-button:hover {
+  background-color: var(--color-primary-hover);
+}
+
+.account-notifications-save-button-saved {
+  background-color: var(--color-primary-subtle);
+  color: var(--color-primary);
+}
+
+.account-notifications-save-button-saved:hover {
+  background-color: var(--color-primary-subtle);
+  color: var(--color-primary);
+}
